fix: require read permissions for git-annex via ssh
The change in 5816106de51046211c8915960c40c82033c13588 made a bug
visible with which it was possible to read annexed content from a
private repository without having read access to it.

Prior to that commit this bug could only be triggered by having a git
version <2.29. I'd expect everyone to use a newer version than that for
a while now.

The fix is to require read permissions at a minimum for git-annex-shell
operations.