fix: git-annex autoinit for pre-created public repositories (#94)
Previously, pre-created public repositories were not
git-annex-initialized on first access via p2phttp. The issue was that
the /config endpoint only runs `git annex init` if the user making the
request has write permissions for the repository. But git-annex first
makes a request without credentials, so to have an authenticated user
the /config endpoint must first respond with a 401. In the case of
pre-created public repositories git-annex instead got a 200 with the
plain-git repository config and never made that second authenticated
request which would have initialized the repository.

Fix this by making the /config endpoint of plain-git repositories
respond with a 401 to unauthenticated requests by a git-annex user
agent.

Fixes #93.

Reviewed-on: https://codeberg.org/forgejo-aneksajo/forgejo-aneksajo/pulls/94
Co-authored-by: Matthias Riße <m.risse@fz-juelich.de>
Co-committed-by: Matthias Riße <m.risse@fz-juelich.de>