Git-annex template repositories (#95)
This adds support for git-annex repositories as templates. A repository
created from a template that is git-annex-initialized will itself be
git-annex-initialized. UUIDs are newly generated such that no clash
happens.

Additionally templates get a new usable variable which contains a
randomly generated UUIDv4. This UUID can e.g. be used in place of the
DataLad dataset ID.

Fixes #22, fixes #30.

Reviewed-on: https://codeberg.org/forgejo-aneksajo/forgejo-aneksajo/pulls/95
Co-authored-by: Matthias Riße <m.risse@fz-juelich.de>
Co-committed-by: Matthias Riße <m.risse@fz-juelich.de>